hydra

+++ src/tinyedm/solver.py
@@ -35,7 +35,9 @@ class DeterministicSolver:
         x1 = x0.to(self.dtype) * self.t_steps[0]
         for i, (t0, t1) in enumerate(zip(self.t_steps[:-1], self.t_steps[1:])):
             x0 = x1
-            denoised = model(x0.to(model.dtype), t0.to(model.dtype)).to(self.dtype)    # mixed precision
+            denoised = model(x0.to(model.dtype), t0.to(model.dtype)).to(
+                self.dtype
+            )  # mixed precision
             dx = (x0 - denoised) / t0
             x1 = x0 + (t1 - t0) * dx
 
@@ -44,4 +46,4 @@ class DeterministicSolver:
                 dx_prime = (x1 - denoised) / t1
                 x1 = x0 + (t1 - t0) * (0.5 * dx + 0.5 * dx_prime)
 
-        return x1.to(x0.dtype) 
+        return x1.to(x0.dtype)
